@@ -82,7 +82,7 @@ async def test_ota_enabled_legacy(tmp_path: pathlib.Path) -> None:
     )
 
     # All are enabled
-    assert len(ota._providers) == 10
+    assert len(ota._providers) == 9
 
 
 async def test_ota_config(tmp_path: pathlib.Path) -> None:
@@ -152,6 +152,8 @@ async def test_ota_config_invalid_provider(tmp_path: pathlib.Path) -> None:
 
 async def test_ota_config_complex(tmp_path: pathlib.Path) -> None:
     # Enable all the providers
+    (tmp_path / "index.json").write_text("{}")
+
     ota = zigpy.ota.OTA(
         config=config.SCHEMA_OTA(
             {
@@ -182,6 +184,19 @@ async def test_ota_config_complex(tmp_path: pathlib.Path) -> None:
                         config.CONF_OTA_PROVIDER_URL: "https://ikea3.example.org/",
                         config.CONF_OTA_PROVIDER_MANUF_IDS: [0xABCD, 0xDCBA],
                     },
+                    {
+                        config.CONF_OTA_PROVIDER_TYPE: "advanced",
+                        config.CONF_OTA_PROVIDER_PATH: tmp_path,
+                        config.CONF_OTA_PROVIDER_WARNING: config.CONF_OTA_ALLOW_ADVANCED_DIR_STRING,
+                    },
+                    {
+                        config.CONF_OTA_PROVIDER_TYPE: "z2m_local",
+                        config.CONF_OTA_PROVIDER_INDEX_FILE: tmp_path / "index.json",
+                    },
+                    {
+                        config.CONF_OTA_PROVIDER_TYPE: "zigpy_local",
+                        config.CONF_OTA_PROVIDER_INDEX_FILE: tmp_path / "index.json",
+                    },
                 ],
             }
         ),
@@ -198,6 +213,9 @@ async def test_ota_config_complex(tmp_path: pathlib.Path) -> None:
             url="https://ikea3.example.org/",
             manufacturer_ids=[0xABCD, 0xDCBA],
         ),
+        zigpy.ota.providers.AdvancedFileProvider(path=tmp_path),
+        zigpy.ota.providers.LocalZ2MProvider(index_file=tmp_path / "index.json"),
+        zigpy.ota.providers.LocalZigpyProvider(index_file=tmp_path / "index.json"),
     ]
 
 
