@@ -32,7 +32,7 @@ from .conftest import (
 )
 
 
-@pytest.fixture()
+@pytest.fixture
 def ieee():
     return make_ieee()
 
@@ -464,9 +464,10 @@ async def test_get_device(app):
 async def test_probe_success():
     config = {"path": "/dev/test"}
 
-    with patch.object(App, "connect") as connect, patch.object(
-        App, "disconnect"
-    ) as disconnect:
+    with (
+        patch.object(App, "connect") as connect,
+        patch.object(App, "disconnect") as disconnect,
+    ):
         result = await App.probe(config)
 
     assert set(config.items()) <= set(result.items())
@@ -478,9 +479,10 @@ async def test_probe_success():
 async def test_probe_failure():
     config = {"path": "/dev/test"}
 
-    with patch.object(
-        App, "connect", side_effect=asyncio.TimeoutError
-    ) as connect, patch.object(App, "disconnect") as disconnect:
+    with (
+        patch.object(App, "connect", side_effect=asyncio.TimeoutError) as connect,
+        patch.object(App, "disconnect") as disconnect,
+    ):
         result = await App.probe(config)
 
     assert result is False
@@ -734,7 +736,7 @@ async def test_request_concurrency():
     assert peak_concurrency == 16
 
 
-@pytest.fixture()
+@pytest.fixture
 def device():
     device = MagicMock()
     device.nwk = 0xABCD
@@ -743,7 +745,7 @@ def device():
     return device
 
 
-@pytest.fixture()
+@pytest.fixture
 def packet(app, device):
     return t.ZigbeePacket(
         src=t.AddrModeAddress(
@@ -889,7 +891,7 @@ async def test_send_broadcast(app, packet):
     )
 
 
-@pytest.fixture()
+@pytest.fixture
 def zdo_packet(app, device):
     return t.ZigbeePacket(
         src=t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=device.nwk),
