@@ -11,7 +11,7 @@ import zigpy.types as t
 import zigpy.zdo.types as zdo_t
 
 
-@pytest.fixture()
+@pytest.fixture
 def backup_factory():
     def inner():
         return zigpy.backups.NetworkBackup(
@@ -92,12 +92,12 @@ def backup_factory():
     return inner
 
 
-@pytest.fixture()
+@pytest.fixture
 def backup(backup_factory):
     return backup_factory()
 
 
-@pytest.fixture()
+@pytest.fixture
 def z2m_backup_json():
     return {
         "metadata": {
@@ -169,7 +169,7 @@ def z2m_backup_json():
     }
 
 
-@pytest.fixture()
+@pytest.fixture
 def zigate_backup_json():
     return {
         "backup_time": "2022-07-20T17:58:16.694438+00:00",
