@@ -24,13 +24,14 @@ import zigpy.types as t
 from zigpy.types.basic import uint16_t
 import zigpy.zcl
 from zigpy.zcl import foundation
+from zigpy.zdo import ZDO
 
 if typing.TYPE_CHECKING:
     from zigpy.application import ControllerApplication
 
 _LOGGER = logging.getLogger(__name__)
 
-_DEVICE_REGISTRY = DeviceRegistry()
+DEVICE_REGISTRY = _DEVICE_REGISTRY = DeviceRegistry()
 _uninitialized_device_message_handlers = []
 
 
@@ -64,17 +65,11 @@ def register_uninitialized_device_message_handler(handler: typing.Callable) -> N
         _uninitialized_device_message_handlers.append(handler)
 
 
-class CustomDevice(zigpy.device.Device):
-    """Implementation of a quirks v1 custom device."""
+class BaseCustomDevice(zigpy.device.Device):
+    """Base class for custom devices."""
 
     _copy_cluster_attr_cache = False
-
     replacement: dict[str, typing.Any] = {}
-    signature = None
-
-    def __init_subclass__(cls) -> None:
-        if getattr(cls, "signature", None) is not None:
-            _DEVICE_REGISTRY.add_to_registry(cls)
 
     def __init__(
         self,
@@ -120,6 +115,36 @@ class CustomDevice(zigpy.device.Device):
         ep = custom_ep_type(self, endpoint_id, replacement_data, replace_device)
         self.endpoints[endpoint_id] = ep
         return ep
+
+    async def apply_custom_configuration(self, *args, **kwargs):
+        """Hook for applications to instruct instances to apply custom configuration."""
+        for endpoint in self.endpoints.values():
+            if isinstance(endpoint, ZDO):
+                continue
+            for cluster in endpoint.in_clusters.values():
+                if (
+                    isinstance(cluster, CustomCluster)
+                    and cluster.apply_custom_configuration
+                    != CustomCluster.apply_custom_configuration
+                ):
+                    await cluster.apply_custom_configuration(*args, **kwargs)
+            for cluster in endpoint.out_clusters.values():
+                if (
+                    isinstance(cluster, CustomCluster)
+                    and cluster.apply_custom_configuration
+                    != CustomCluster.apply_custom_configuration
+                ):
+                    await cluster.apply_custom_configuration(*args, **kwargs)
+
+
+class CustomDevice(BaseCustomDevice):
+    """Implementation of a quirks v1 custom device."""
+
+    signature = None
+
+    def __init_subclass__(cls) -> None:
+        if getattr(cls, "signature", None) is not None:
+            _DEVICE_REGISTRY.add_to_registry(cls)
 
 
 class CustomEndpoint(zigpy.endpoint.Endpoint):
@@ -272,13 +297,16 @@ class CustomCluster(zigpy.zcl.Cluster):
 
         succeeded = [
             foundation.ReadAttributeRecord(
-                attr, foundation.Status.SUCCESS, foundation.TypeValue()
+                attrid=attr,
+                status=foundation.Status.SUCCESS,
+                value=foundation.TypeValue(
+                    type=None,
+                    value=self._CONSTANT_ATTRIBUTES[attr],
+                ),
             )
             for attr in attributes
             if attr in self._CONSTANT_ATTRIBUTES
         ]
-        for record in succeeded:
-            record.value.value = self._CONSTANT_ATTRIBUTES[record.attrid]
 
         attrs_to_read = [
             attr for attr in attributes if attr not in self._CONSTANT_ATTRIBUTES
