@@ -9,6 +9,7 @@ from zigpy.zcl import Cluster, foundation
 from zigpy.zcl.foundation import (
     BaseAttributeDefs,
     BaseCommandDefs,
+    Direction,
     ZCLAttributeDef,
     ZCLCommandDef,
 )
@@ -407,21 +408,29 @@ class DoorLock(Cluster):
 
     class ServerCommandDefs(BaseCommandDefs):
         lock_door: Final = ZCLCommandDef(
-            id=0x00, schema={"pin_code?": t.CharacterString}, direction=False
+            id=0x00,
+            schema={"pin_code?": t.CharacterString},
+            direction=Direction.Client_to_Server,
         )
         unlock_door: Final = ZCLCommandDef(
-            id=0x01, schema={"pin_code?": t.CharacterString}, direction=False
+            id=0x01,
+            schema={"pin_code?": t.CharacterString},
+            direction=Direction.Client_to_Server,
         )
         toggle_door: Final = ZCLCommandDef(
-            id=0x02, schema={"pin_code?": t.CharacterString}, direction=False
+            id=0x02,
+            schema={"pin_code?": t.CharacterString},
+            direction=Direction.Client_to_Server,
         )
         unlock_with_timeout: Final = ZCLCommandDef(
             id=0x03,
             schema={"timeout": t.uint16_t, "pin_code?": t.CharacterString},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_log_record: Final = ZCLCommandDef(
-            id=0x04, schema={"log_index": t.uint16_t}, direction=False
+            id=0x04,
+            schema={"log_index": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
         set_pin_code: Final = ZCLCommandDef(
             id=0x05,
@@ -431,22 +440,30 @@ class DoorLock(Cluster):
                 "user_type": UserType,
                 "pin_code": t.CharacterString,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_pin_code: Final = ZCLCommandDef(
-            id=0x06, schema={"user_id": t.uint16_t}, direction=False
+            id=0x06,
+            schema={"user_id": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
         clear_pin_code: Final = ZCLCommandDef(
-            id=0x07, schema={"user_id": t.uint16_t}, direction=False
+            id=0x07,
+            schema={"user_id": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
-        clear_all_pin_codes: Final = ZCLCommandDef(id=0x08, schema={}, direction=False)
+        clear_all_pin_codes: Final = ZCLCommandDef(
+            id=0x08, schema={}, direction=Direction.Client_to_Server
+        )
         set_user_status: Final = ZCLCommandDef(
             id=0x09,
             schema={"user_id": t.uint16_t, "user_status": UserStatus},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_user_status: Final = ZCLCommandDef(
-            id=0x0A, schema={"user_id": t.uint16_t}, direction=False
+            id=0x0A,
+            schema={"user_id": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
         set_week_day_schedule: Final = ZCLCommandDef(
             id=0x0B,
@@ -459,17 +476,17 @@ class DoorLock(Cluster):
                 "end_hour": t.uint8_t,
                 "end_minute": t.uint8_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_week_day_schedule: Final = ZCLCommandDef(
             id=0x0C,
             schema={"schedule_id": t.uint8_t, "user_id": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         clear_week_day_schedule: Final = ZCLCommandDef(
             id=0x0D,
             schema={"schedule_id": t.uint8_t, "user_id": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         set_year_day_schedule: Final = ZCLCommandDef(
             id=0x0E,
@@ -479,17 +496,17 @@ class DoorLock(Cluster):
                 "local_start_time": t.LocalTime,
                 "local_end_time": t.LocalTime,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_year_day_schedule: Final = ZCLCommandDef(
             id=0x0F,
             schema={"schedule_id": t.uint8_t, "user_id": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         clear_year_day_schedule: Final = ZCLCommandDef(
             id=0x10,
             schema={"schedule_id": t.uint8_t, "user_id": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         set_holiday_schedule: Final = ZCLCommandDef(
             id=0x11,
@@ -499,21 +516,27 @@ class DoorLock(Cluster):
                 "local_end_time": t.LocalTime,
                 "operating_mode_during_holiday": OperatingMode,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_holiday_schedule: Final = ZCLCommandDef(
-            id=0x12, schema={"holiday_schedule_id": t.uint8_t}, direction=False
+            id=0x12,
+            schema={"holiday_schedule_id": t.uint8_t},
+            direction=Direction.Client_to_Server,
         )
         clear_holiday_schedule: Final = ZCLCommandDef(
-            id=0x13, schema={"holiday_schedule_id": t.uint8_t}, direction=False
+            id=0x13,
+            schema={"holiday_schedule_id": t.uint8_t},
+            direction=Direction.Client_to_Server,
         )
         set_user_type: Final = ZCLCommandDef(
             id=0x14,
             schema={"user_id": t.uint16_t, "user_type": UserType},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_user_type: Final = ZCLCommandDef(
-            id=0x15, schema={"user_id": t.uint16_t}, direction=False
+            id=0x15,
+            schema={"user_id": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
         set_rfid_code: Final = ZCLCommandDef(
             id=0x16,
@@ -523,28 +546,42 @@ class DoorLock(Cluster):
                 "user_type": UserType,
                 "rfid_code": t.CharacterString,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_rfid_code: Final = ZCLCommandDef(
-            id=0x17, schema={"user_id": t.uint16_t}, direction=False
+            id=0x17,
+            schema={"user_id": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
         clear_rfid_code: Final = ZCLCommandDef(
-            id=0x18, schema={"user_id": t.uint16_t}, direction=False
+            id=0x18,
+            schema={"user_id": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
-        clear_all_rfid_codes: Final = ZCLCommandDef(id=0x19, schema={}, direction=False)
+        clear_all_rfid_codes: Final = ZCLCommandDef(
+            id=0x19, schema={}, direction=Direction.Client_to_Server
+        )
 
     class ClientCommandDefs(BaseCommandDefs):
         lock_door_response: Final = ZCLCommandDef(
-            id=0x00, schema={"status": foundation.Status}, direction=True
+            id=0x00,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         unlock_door_response: Final = ZCLCommandDef(
-            id=0x01, schema={"status": foundation.Status}, direction=True
+            id=0x01,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         toggle_door_response: Final = ZCLCommandDef(
-            id=0x02, schema={"status": foundation.Status}, direction=True
+            id=0x02,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         unlock_with_timeout_response: Final = ZCLCommandDef(
-            id=0x03, schema={"status": foundation.Status}, direction=True
+            id=0x03,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_log_record_response: Final = ZCLCommandDef(
             id=0x04,
@@ -557,10 +594,12 @@ class DoorLock(Cluster):
                 "user_id": t.uint16_t,
                 "pin?": t.CharacterString,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         set_pin_code_response: Final = ZCLCommandDef(
-            id=0x05, schema={"status": foundation.Status}, direction=True
+            id=0x05,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_pin_code_response: Final = ZCLCommandDef(
             id=0x06,
@@ -570,24 +609,32 @@ class DoorLock(Cluster):
                 "user_type": UserType,
                 "code": t.CharacterString,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         clear_pin_code_response: Final = ZCLCommandDef(
-            id=0x07, schema={"status": foundation.Status}, direction=True
+            id=0x07,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         clear_all_pin_codes_response: Final = ZCLCommandDef(
-            id=0x08, schema={"status": foundation.Status}, direction=True
+            id=0x08,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         set_user_status_response: Final = ZCLCommandDef(
-            id=0x09, schema={"status": foundation.Status}, direction=True
+            id=0x09,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_user_status_response: Final = ZCLCommandDef(
             id=0x0A,
             schema={"user_id": t.uint16_t, "user_status": UserStatus},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         set_week_day_schedule_response: Final = ZCLCommandDef(
-            id=0x0B, schema={"status": foundation.Status}, direction=True
+            id=0x0B,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_week_day_schedule_response: Final = ZCLCommandDef(
             id=0x0C,
@@ -601,13 +648,17 @@ class DoorLock(Cluster):
                 "end_hour?": t.uint8_t,
                 "end_minute?": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         clear_week_day_schedule_response: Final = ZCLCommandDef(
-            id=0x0D, schema={"status": foundation.Status}, direction=True
+            id=0x0D,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         set_year_day_schedule_response: Final = ZCLCommandDef(
-            id=0x0E, schema={"status": foundation.Status}, direction=True
+            id=0x0E,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_year_day_schedule_response: Final = ZCLCommandDef(
             id=0x0F,
@@ -618,13 +669,17 @@ class DoorLock(Cluster):
                 "local_start_time?": t.LocalTime,
                 "local_end_time?": t.LocalTime,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         clear_year_day_schedule_response: Final = ZCLCommandDef(
-            id=0x10, schema={"status": foundation.Status}, direction=True
+            id=0x10,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         set_holiday_schedule_response: Final = ZCLCommandDef(
-            id=0x11, schema={"status": foundation.Status}, direction=True
+            id=0x11,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_holiday_schedule_response: Final = ZCLCommandDef(
             id=0x12,
@@ -635,21 +690,27 @@ class DoorLock(Cluster):
                 "local_end_time?": t.LocalTime,
                 "operating_mode_during_holiday?": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         clear_holiday_schedule_response: Final = ZCLCommandDef(
-            id=0x13, schema={"status": foundation.Status}, direction=True
+            id=0x13,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         set_user_type_response: Final = ZCLCommandDef(
-            id=0x14, schema={"status": foundation.Status}, direction=True
+            id=0x14,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_user_type_response: Final = ZCLCommandDef(
             id=0x15,
             schema={"user_id": t.uint16_t, "user_type": UserType},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         set_rfid_code_response: Final = ZCLCommandDef(
-            id=0x16, schema={"status": foundation.Status}, direction=True
+            id=0x16,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         get_rfid_code_response: Final = ZCLCommandDef(
             id=0x17,
@@ -659,13 +720,17 @@ class DoorLock(Cluster):
                 "user_type": UserType,
                 "rfid_code": t.CharacterString,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         clear_rfid_code_response: Final = ZCLCommandDef(
-            id=0x18, schema={"status": foundation.Status}, direction=True
+            id=0x18,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         clear_all_rfid_codes_response: Final = ZCLCommandDef(
-            id=0x19, schema={"status": foundation.Status}, direction=True
+            id=0x19,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         operation_event_notification: Final = ZCLCommandDef(
             id=0x20,
@@ -677,7 +742,7 @@ class DoorLock(Cluster):
                 "local_time": t.LocalTime,
                 "data?": t.CharacterString,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         programming_event_notification: Final = ZCLCommandDef(
             id=0x21,
@@ -691,7 +756,7 @@ class DoorLock(Cluster):
                 "local_time": t.LocalTime,
                 "data?": t.CharacterString,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
 
 
@@ -798,18 +863,92 @@ class WindowCovering(Cluster):
         )
 
     class ServerCommandDefs(BaseCommandDefs):
-        up_open: Final = ZCLCommandDef(id=0x00, schema={}, direction=False)
-        down_close: Final = ZCLCommandDef(id=0x01, schema={}, direction=False)
-        stop: Final = ZCLCommandDef(id=0x02, schema={}, direction=False)
+        up_open: Final = ZCLCommandDef(
+            id=0x00, schema={}, direction=Direction.Client_to_Server
+        )
+        down_close: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
+        stop: Final = ZCLCommandDef(
+            id=0x02, schema={}, direction=Direction.Client_to_Server
+        )
         go_to_lift_value: Final = ZCLCommandDef(
-            id=0x04, schema={"lift_value": t.uint16_t}, direction=False
+            id=0x04,
+            schema={"lift_value": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
         go_to_lift_percentage: Final = ZCLCommandDef(
-            id=0x05, schema={"percentage_lift_value": t.uint8_t}, direction=False
+            id=0x05,
+            schema={"percentage_lift_value": t.uint8_t},
+            direction=Direction.Client_to_Server,
         )
         go_to_tilt_value: Final = ZCLCommandDef(
-            id=0x07, schema={"tilt_value": t.uint16_t}, direction=False
+            id=0x07,
+            schema={"tilt_value": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
         go_to_tilt_percentage: Final = ZCLCommandDef(
-            id=0x08, schema={"percentage_tilt_value": t.uint8_t}, direction=False
+            id=0x08,
+            schema={"percentage_tilt_value": t.uint8_t},
+            direction=Direction.Client_to_Server,
         )
+
+
+class MovingState(t.enum8):
+    Stopped = 0x00
+    Closing = 0x01
+    Opening = 0x02
+
+
+class SafetyStatus(t.bitmap16):
+    Remote_Lockout = 0b00000000_00000001
+    Tamper_Detected = 0b00000000_00000010
+    Failed_Communication = 0b00000000_00000100
+    Position_Failure = 0b00000000_00001000
+
+
+class Capabilities(t.bitmap8):
+    Partial_Barrier = 0b00000001
+
+
+class BarrierControl(Cluster):
+    cluster_id: Final = 0x0103
+    name: Final = "Barrier Control"
+    ep_attribute: Final = "barrier_control"
+
+    class AttributeDefs(BaseAttributeDefs):
+        moving_state: Final = ZCLAttributeDef(
+            id=0x0001, type=MovingState, access="rp", mandatory=True
+        )
+        safety_status: Final = ZCLAttributeDef(
+            id=0x0002, type=SafetyStatus, access="rp", mandatory=True
+        )
+        capabilities: Final = ZCLAttributeDef(
+            id=0x0003, type=Capabilities, access="r", mandatory=True
+        )
+        open_events: Final = ZCLAttributeDef(id=0x0004, type=t.uint16_t, access="rw")
+        close_events: Final = ZCLAttributeDef(id=0x0005, type=t.uint16_t, access="rw")
+        command_open_events: Final = ZCLAttributeDef(
+            id=0x0006, type=t.uint16_t, access="rw"
+        )
+        command_close_events: Final = ZCLAttributeDef(
+            id=0x0007, type=t.uint16_t, access="rw"
+        )
+        open_period: Final = ZCLAttributeDef(id=0x0008, type=t.uint16_t, access="rw")
+        close_period: Final = ZCLAttributeDef(id=0x0009, type=t.uint16_t, access="rw")
+        barrier_position: Final = ZCLAttributeDef(
+            id=0x000A, type=t.uint8_t, access="rps", mandatory=True
+        )
+
+    class ServerCommandDefs(BaseCommandDefs):
+        go_to_percent: Final = ZCLCommandDef(
+            id=0x00,
+            schema={"percent_open": t.uint8_t},
+            direction=Direction.Client_to_Server,
+        )
+        stop: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
+
+    class ClientCommandDefs(BaseCommandDefs):
+        pass
