@@ -11,9 +11,12 @@ from zigpy.zcl import Cluster, foundation
 from zigpy.zcl.foundation import (
     BaseAttributeDefs,
     BaseCommandDefs,
+    Direction,
     ZCLAttributeDef,
     ZCLCommandDef,
 )
+
+ZIGBEE_EPOCH = datetime(2000, 1, 1, 0, 0, 0, 0, tzinfo=timezone.utc)
 
 
 class PowerSource(t.enum8):
@@ -276,7 +279,9 @@ class Basic(Cluster):
         reporting_status: Final = foundation.ZCL_REPORTING_STATUS_ATTR
 
     class ServerCommandDefs(BaseCommandDefs):
-        reset_fact_default: Final = ZCLCommandDef(id=0x00, schema={}, direction=False)
+        reset_fact_default: Final = ZCLCommandDef(
+            id=0x00, schema={}, direction=Direction.Client_to_Server
+        )
 
 
 class MainsAlarmMask(t.bitmap8):
@@ -569,20 +574,26 @@ class Identify(Cluster):
 
     class ServerCommandDefs(BaseCommandDefs):
         identify: Final = ZCLCommandDef(
-            id=0x00, schema={"identify_time": t.uint16_t}, direction=False
+            id=0x00,
+            schema={"identify_time": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
-        identify_query: Final = ZCLCommandDef(id=0x01, schema={}, direction=False)
+        identify_query: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
         # 0x02: ("ezmode_invoke", (t.bitmap8,), False),
         # 0x03: ("update_commission_state", (t.bitmap8,), False),
         trigger_effect: Final = ZCLCommandDef(
             id=0x40,
             schema={"effect_id": EffectIdentifier, "effect_variant": EffectVariant},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
         identify_query_response: Final = ZCLCommandDef(
-            id=0x00, schema={"timeout": t.uint16_t}, direction=True
+            id=0x00,
+            schema={"timeout": t.uint16_t},
+            direction=Direction.Server_to_Client,
         )
 
 
@@ -611,29 +622,33 @@ class Groups(Cluster):
         add: Final = ZCLCommandDef(
             id=0x00,
             schema={"group_id": t.Group, "group_name": t.LimitedCharString(16)},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         view: Final = ZCLCommandDef(
-            id=0x01, schema={"group_id": t.Group}, direction=False
+            id=0x01, schema={"group_id": t.Group}, direction=Direction.Client_to_Server
         )
         get_membership: Final = ZCLCommandDef(
-            id=0x02, schema={"groups": t.LVList[t.Group]}, direction=False
+            id=0x02,
+            schema={"groups": t.LVList[t.Group]},
+            direction=Direction.Client_to_Server,
         )
         remove: Final = ZCLCommandDef(
-            id=0x03, schema={"group_id": t.Group}, direction=False
+            id=0x03, schema={"group_id": t.Group}, direction=Direction.Client_to_Server
         )
-        remove_all: Final = ZCLCommandDef(id=0x04, schema={}, direction=False)
+        remove_all: Final = ZCLCommandDef(
+            id=0x04, schema={}, direction=Direction.Client_to_Server
+        )
         add_if_identifying: Final = ZCLCommandDef(
             id=0x05,
             schema={"group_id": t.Group, "group_name": t.LimitedCharString(16)},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
         add_response: Final = ZCLCommandDef(
             id=0x00,
             schema={"status": foundation.Status, "group_id": t.Group},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         view_response: Final = ZCLCommandDef(
             id=0x01,
@@ -642,17 +657,17 @@ class Groups(Cluster):
                 "group_id": t.Group,
                 "group_name": t.LimitedCharString(16),
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         get_membership_response: Final = ZCLCommandDef(
             id=0x02,
             schema={"capacity": t.uint8_t, "groups": t.LVList[t.Group]},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         remove_response: Final = ZCLCommandDef(
             id=0x03,
             schema={"status": foundation.Status, "group_id": t.Group},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
 
 
@@ -696,26 +711,26 @@ class Scenes(Cluster):
                 "transition_time": t.uint16_t,
                 "scene_name": t.LimitedCharString(16),
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         # TODO: + extension field sets
         view: Final = ZCLCommandDef(
             id=0x01,
             schema={"group_id": t.Group, "scene_id": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         remove: Final = ZCLCommandDef(
             id=0x02,
             schema={"group_id": t.Group, "scene_id": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         remove_all: Final = ZCLCommandDef(
-            id=0x03, schema={"group_id": t.Group}, direction=False
+            id=0x03, schema={"group_id": t.Group}, direction=Direction.Client_to_Server
         )
         store: Final = ZCLCommandDef(
             id=0x04,
             schema={"group_id": t.Group, "scene_id": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         recall: Final = ZCLCommandDef(
             id=0x05,
@@ -724,10 +739,10 @@ class Scenes(Cluster):
                 "scene_id": t.uint8_t,
                 "transition_time?": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_scene_membership: Final = ZCLCommandDef(
-            id=0x06, schema={"group_id": t.Group}, direction=False
+            id=0x06, schema={"group_id": t.Group}, direction=Direction.Client_to_Server
         )
         enhanced_add: Final = ZCLCommandDef(
             id=0x40,
@@ -737,12 +752,12 @@ class Scenes(Cluster):
                 "transition_time": t.uint16_t,
                 "scene_name": t.LimitedCharString(16),
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         enhanced_view: Final = ZCLCommandDef(
             id=0x41,
             schema={"group_id": t.Group, "scene_id": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         copy: Final = ZCLCommandDef(
             id=0x42,
@@ -753,7 +768,7 @@ class Scenes(Cluster):
                 "group_id_to": t.uint16_t,
                 "scene_id_to": t.uint8_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
@@ -764,7 +779,7 @@ class Scenes(Cluster):
                 "group_id": t.Group,
                 "scene_id": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         view_response: Final = ZCLCommandDef(
             id=0x01,
@@ -775,7 +790,7 @@ class Scenes(Cluster):
                 "transition_time?": t.uint16_t,
                 "scene_name?": t.LimitedCharString(16),
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         # TODO: + extension field sets
         remove_scene_response: Final = ZCLCommandDef(
@@ -785,12 +800,12 @@ class Scenes(Cluster):
                 "group_id": t.Group,
                 "scene_id": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         remove_all_scenes_response: Final = ZCLCommandDef(
             id=0x03,
             schema={"status": foundation.Status, "group_id": t.Group},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         store_scene_response: Final = ZCLCommandDef(
             id=0x04,
@@ -799,7 +814,7 @@ class Scenes(Cluster):
                 "group_id": t.Group,
                 "scene_id": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         get_scene_membership_response: Final = ZCLCommandDef(
             id=0x06,
@@ -809,7 +824,7 @@ class Scenes(Cluster):
                 "group_id": t.Group,
                 "scenes?": t.LVList[t.uint8_t],
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         enhanced_add_response: Final = ZCLCommandDef(
             id=0x40,
@@ -818,7 +833,7 @@ class Scenes(Cluster):
                 "group_id": t.Group,
                 "scene_id": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         enhanced_view_response: Final = ZCLCommandDef(
             id=0x41,
@@ -829,7 +844,7 @@ class Scenes(Cluster):
                 "transition_time?": t.uint16_t,
                 "scene_name?": t.LimitedCharString(16),
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         # TODO: + extension field sets
         copy_response: Final = ZCLCommandDef(
@@ -839,7 +854,7 @@ class Scenes(Cluster):
                 "group_id": t.Group,
                 "scene_id": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
 
 
@@ -894,16 +909,22 @@ class OnOff(Cluster):
         reporting_status: Final = foundation.ZCL_REPORTING_STATUS_ATTR
 
     class ServerCommandDefs(BaseCommandDefs):
-        off: Final = ZCLCommandDef(id=0x00, schema={}, direction=False)
-        on: Final = ZCLCommandDef(id=0x01, schema={}, direction=False)
-        toggle: Final = ZCLCommandDef(id=0x02, schema={}, direction=False)
+        off: Final = ZCLCommandDef(
+            id=0x00, schema={}, direction=Direction.Client_to_Server
+        )
+        on: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
+        toggle: Final = ZCLCommandDef(
+            id=0x02, schema={}, direction=Direction.Client_to_Server
+        )
         off_with_effect: Final = ZCLCommandDef(
             id=0x40,
             schema={"effect_id": OffEffectIdentifier, "effect_variant": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         on_with_recall_global_scene: Final = ZCLCommandDef(
-            id=0x41, schema={}, direction=False
+            id=0x41, schema={}, direction=Direction.Client_to_Server
         )
         on_with_timed_off: Final = ZCLCommandDef(
             id=0x42,
@@ -912,7 +933,7 @@ class OnOff(Cluster):
                 "on_time": t.uint16_t,
                 "off_wait_time": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
 
@@ -959,6 +980,11 @@ class StepMode(t.enum8):
     Down = 0x01
 
 
+class OptionsMask(t.bitmap8):
+    Execute_if_off_present = 0b00000001
+    Couple_color_temp_to_level_present = 0b00000010
+
+
 class Options(t.bitmap8):
     Execute_if_off = 0b00000001
     Couple_color_temp_to_level = 0b00000010
@@ -972,6 +998,7 @@ class LevelControl(Cluster):
     MoveMode: Final = MoveMode
     StepMode: Final = StepMode
     Options: Final = Options
+    OptionsMask: Final = OptionsMask
 
     cluster_id: Final[t.uint16_t] = 0x0008
     name: Final = "Level control"
@@ -1015,20 +1042,20 @@ class LevelControl(Cluster):
             schema={
                 "level": t.uint8_t,
                 "transition_time": t.uint16_t,
-                "options_mask?": t.bitmap8,
-                "options_override?": t.bitmap8,
+                "options_mask?": OptionsMask,
+                "options_override?": Options,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         move: Final = ZCLCommandDef(
             id=0x01,
             schema={
                 "move_mode": MoveMode,
                 "rate": t.uint8_t,
-                "options_mask?": t.bitmap8,
-                "options_override?": t.bitmap8,
+                "options_mask?": OptionsMask,
+                "options_override?": Options,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         step: Final = ZCLCommandDef(
             id=0x02,
@@ -1036,28 +1063,28 @@ class LevelControl(Cluster):
                 "step_mode": StepMode,
                 "step_size": t.uint8_t,
                 "transition_time": t.uint16_t,
-                "options_mask?": t.bitmap8,
-                "options_override?": t.bitmap8,
+                "options_mask?": OptionsMask,
+                "options_override?": Options,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         stop: Final = ZCLCommandDef(
             id=0x03,
             schema={
-                "options_mask?": t.bitmap8,
-                "options_override?": t.bitmap8,
+                "options_mask?": OptionsMask,
+                "options_override?": Options,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         move_to_level_with_on_off: Final = ZCLCommandDef(
             id=0x04,
             schema={"level": t.uint8_t, "transition_time": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         move_with_on_off: Final = ZCLCommandDef(
             id=0x05,
             schema={"move_mode": MoveMode, "rate": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         step_with_on_off: Final = ZCLCommandDef(
             id=0x06,
@@ -1066,11 +1093,15 @@ class LevelControl(Cluster):
                 "step_size": t.uint8_t,
                 "transition_time": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
-        stop_with_on_off: Final = ZCLCommandDef(id=0x07, schema={}, direction=False)
+        stop_with_on_off: Final = ZCLCommandDef(
+            id=0x07, schema={}, direction=Direction.Client_to_Server
+        )
         move_to_closest_frequency: Final = ZCLCommandDef(
-            id=0x08, schema={"frequency": t.uint16_t}, direction=False
+            id=0x08,
+            schema={"frequency": t.uint16_t},
+            direction=Direction.Client_to_Server,
         )
 
 
@@ -1091,18 +1122,24 @@ class Alarms(Cluster):
         reset_alarm: Final = ZCLCommandDef(
             id=0x00,
             schema={"alarm_code": t.uint8_t, "cluster_id": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
-        reset_all_alarms: Final = ZCLCommandDef(id=0x01, schema={}, direction=False)
-        get_alarm: Final = ZCLCommandDef(id=0x02, schema={}, direction=False)
-        reset_alarm_log: Final = ZCLCommandDef(id=0x03, schema={}, direction=False)
+        reset_all_alarms: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
+        get_alarm: Final = ZCLCommandDef(
+            id=0x02, schema={}, direction=Direction.Client_to_Server
+        )
+        reset_alarm_log: Final = ZCLCommandDef(
+            id=0x03, schema={}, direction=Direction.Client_to_Server
+        )
         # 0x04: ("publish_event_log", {}, False),
 
     class ClientCommandDefs(BaseCommandDefs):
         alarm: Final = ZCLCommandDef(
             id=0x00,
             schema={"alarm_code": t.uint8_t, "cluster_id": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_alarm_response: Final = ZCLCommandDef(
             id=0x01,
@@ -1112,7 +1149,7 @@ class Alarms(Cluster):
                 "cluster_id?": t.uint16_t,
                 "timestamp?": t.uint32_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         # 0x02: ("get_event_log", {}, False),
 
@@ -1139,7 +1176,7 @@ class Time(Cluster):
             id=0x0000, type=t.UTCTime, access="r*w", mandatory=True
         )
         time_status: Final = ZCLAttributeDef(
-            id=0x0001, type=t.bitmap8, access="r*w", mandatory=True
+            id=0x0001, type=TimeStatus, access="r*w", mandatory=True
         )
         time_zone: Final = ZCLAttributeDef(id=0x0002, type=t.int32s, access="rw")
         dst_start: Final = ZCLAttributeDef(id=0x0003, type=t.uint32_t, access="rw")
@@ -1161,27 +1198,55 @@ class Time(Cluster):
         hdr: foundation.ZCLHeader,
         *args: list[Any],
         dst_addressing: AddressingMode | None = None,
-    ):
+    ) -> None:
+        super().handle_cluster_general_request(hdr, args, dst_addressing=dst_addressing)
+
         if hdr.command_id == foundation.GeneralCommand.Read_Attributes:
-            data = {}
-            for attr in args[0][0]:
-                if attr == 0:
-                    epoch = datetime(2000, 1, 1, 0, 0, 0, 0, tzinfo=timezone.utc)
-                    diff = datetime.now(timezone.utc) - epoch
-                    data[attr] = diff.total_seconds()
-                elif attr == 1:
-                    data[attr] = 7
-                elif attr == 2:
-                    local_time = datetime.fromtimestamp(86400).astimezone()
-                    utc_time = datetime.fromtimestamp(86400, timezone.utc)
-                    data[attr] = (local_time - utc_time).total_seconds()
-                elif attr == 7:
-                    epoch = datetime(2000, 1, 1, 0, 0, 0, 0)
-                    diff = datetime.now() - epoch
-                    data[attr] = diff.total_seconds()
+            now = datetime.now(timezone.utc)
+            tz_offset = datetime.now().astimezone().utcoffset()
+            assert tz_offset is not None
+
+            records = []
+
+            for attr in args[0].attribute_ids:
+                record = foundation.ReadAttributeRecord(attrid=attr)
+
+                if attr == self.AttributeDefs.time.id:
+                    record.status = foundation.Status.SUCCESS
+                    record.value = foundation.TypeValue(
+                        type=self.AttributeDefs.time.zcl_type,
+                        value=t.UTCTime((now - ZIGBEE_EPOCH).total_seconds()),
+                    )
+                elif attr == self.AttributeDefs.time_status.id:
+                    record.status = foundation.Status.SUCCESS
+                    record.value = foundation.TypeValue(
+                        type=self.AttributeDefs.time_status.zcl_type,
+                        value=(
+                            TimeStatus.Master
+                            | TimeStatus.Synchronized
+                            | TimeStatus.Master_for_Zone_and_DST
+                        ),
+                    )
+                elif attr == self.AttributeDefs.time_zone.id:
+                    record.status = foundation.Status.SUCCESS
+                    record.value = foundation.TypeValue(
+                        type=self.AttributeDefs.time_zone.zcl_type,
+                        value=t.int32s(tz_offset.total_seconds()),
+                    )
+                elif attr == self.AttributeDefs.local_time.id:
+                    record.status = foundation.Status.SUCCESS
+                    record.value = foundation.TypeValue(
+                        type=self.AttributeDefs.local_time.zcl_type,
+                        value=t.LocalTime(
+                            (now + tz_offset - ZIGBEE_EPOCH).total_seconds()
+                        ),
+                    )
                 else:
-                    data[attr] = None
-            self.create_catching_task(self.read_attributes_rsp(data, tsn=hdr.tsn))
+                    record.status = foundation.Status.UNSUPPORTED_ATTRIBUTE
+
+                records.append(record)
+
+            self.create_catching_task(self.read_attributes_rsp(records, tsn=hdr.tsn))
 
 
 class LocationMethod(t.enum8):
@@ -1260,7 +1325,7 @@ class RSSILocation(Cluster):
                 "power": t.int16s,
                 "path_loss_exponent": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         set_dev_config: Final = ZCLCommandDef(
             id=0x01,
@@ -1271,10 +1336,12 @@ class RSSILocation(Cluster):
                 "num_rssi_measurements": t.uint8_t,
                 "reporting_period": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_dev_config: Final = ZCLCommandDef(
-            id=0x02, schema={"target_addr": t.EUI64}, direction=False
+            id=0x02,
+            schema={"target_addr": t.EUI64},
+            direction=Direction.Client_to_Server,
         )
         get_location_data: Final = ZCLCommandDef(
             id=0x03,
@@ -1283,7 +1350,7 @@ class RSSILocation(Cluster):
                 "num_responses": t.uint8_t,
                 "target_addr": t.EUI64,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         rssi_response: Final = ZCLCommandDef(
             id=0x04,
@@ -1295,7 +1362,7 @@ class RSSILocation(Cluster):
                 "rssi": t.int8s,
                 "num_rssi_measurements": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         send_pings: Final = ZCLCommandDef(
             id=0x05,
@@ -1304,7 +1371,7 @@ class RSSILocation(Cluster):
                 "num_rssi_measurements": t.uint8_t,
                 "calculation_period": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         anchor_node_announce: Final = ZCLCommandDef(
             id=0x06,
@@ -1314,7 +1381,7 @@ class RSSILocation(Cluster):
                 "y": t.int16s,
                 "z": t.int16s,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
@@ -1328,7 +1395,7 @@ class RSSILocation(Cluster):
                 "num_rssi_measurements?": t.uint8_t,
                 "reporting_period?": t.uint16_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         location_data_response: Final = ZCLCommandDef(
             id=0x01,
@@ -1344,28 +1411,34 @@ class RSSILocation(Cluster):
                 "quality_measure?": t.uint8_t,
                 "location_age?": t.uint16_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         location_data_notification: Final = ZCLCommandDef(
-            id=0x02, schema={}, direction=False
+            id=0x02, schema={}, direction=Direction.Client_to_Server
         )
         compact_location_data_notification: Final = ZCLCommandDef(
-            id=0x03, schema={}, direction=False
+            id=0x03, schema={}, direction=Direction.Client_to_Server
         )
         rssi_ping: Final = ZCLCommandDef(
-            id=0x04, schema={"location_type": t.uint8_t}, direction=False
+            id=0x04,
+            schema={"location_type": t.uint8_t},
+            direction=Direction.Client_to_Server,
         )
-        rssi_req: Final = ZCLCommandDef(id=0x05, schema={}, direction=False)
+        rssi_req: Final = ZCLCommandDef(
+            id=0x05, schema={}, direction=Direction.Client_to_Server
+        )
         report_rssi_measurements: Final = ZCLCommandDef(
             id=0x06,
             schema={
                 "measuring_device": t.EUI64,
                 "neighbors": t.LVList[NeighborInfo],
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         request_own_location: Final = ZCLCommandDef(
-            id=0x07, schema={"ieee_of_blind_node": t.EUI64}, direction=False
+            id=0x07,
+            schema={"ieee_of_blind_node": t.EUI64},
+            direction=Direction.Client_to_Server,
         )
 
 
@@ -1623,7 +1696,7 @@ class MultistateInput(Cluster):
 
     class AttributeDefs(BaseAttributeDefs):
         state_text: Final = ZCLAttributeDef(
-            id=0x000E, type=t.List[t.CharacterString], access="r*w"
+            id=0x000E, type=t.LVList[t.CharacterString, t.uint16_t], access="r*w"
         )
         description: Final = ZCLAttributeDef(
             id=0x001C, type=t.CharacterString, access="r*w"
@@ -1656,7 +1729,7 @@ class MultistateOutput(Cluster):
 
     class AttributeDefs(BaseAttributeDefs):
         state_text: Final = ZCLAttributeDef(
-            id=0x000E, type=t.List[t.CharacterString], access="r*w"
+            id=0x000E, type=t.LVList[t.CharacterString, t.uint16_t], access="r*w"
         )
         description: Final = ZCLAttributeDef(
             id=0x001C, type=t.CharacterString, access="r*w"
@@ -1692,7 +1765,7 @@ class MultistateValue(Cluster):
 
     class AttributeDefs(BaseAttributeDefs):
         state_text: Final = ZCLAttributeDef(
-            id=0x000E, type=t.List[t.CharacterString], access="r*w"
+            id=0x000E, type=t.LVList[t.CharacterString, t.uint16_t], access="r*w"
         )
         description: Final = ZCLAttributeDef(
             id=0x001C, type=t.CharacterString, access="r*w"
@@ -1824,36 +1897,44 @@ class Commissioning(Cluster):
         restart_device: Final = ZCLCommandDef(
             id=0x00,
             schema={"options": t.bitmap8, "delay": t.uint8_t, "jitter": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         save_startup_parameters: Final = ZCLCommandDef(
             id=0x01,
             schema={"options": t.bitmap8, "index": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         restore_startup_parameters: Final = ZCLCommandDef(
             id=0x02,
             schema={"options": t.bitmap8, "index": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         reset_startup_parameters: Final = ZCLCommandDef(
             id=0x03,
             schema={"options": t.bitmap8, "index": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
         restart_device_response: Final = ZCLCommandDef(
-            id=0x00, schema={"status": foundation.Status}, direction=True
+            id=0x00,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         save_startup_params_response: Final = ZCLCommandDef(
-            id=0x01, schema={"status": foundation.Status}, direction=True
+            id=0x01,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         restore_startup_params_response: Final = ZCLCommandDef(
-            id=0x02, schema={"status": foundation.Status}, direction=True
+            id=0x02,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
         reset_startup_params_response: Final = ZCLCommandDef(
-            id=0x03, schema={"status": foundation.Status}, direction=True
+            id=0x03,
+            schema={"status": foundation.Status},
+            direction=Direction.Server_to_Client,
         )
 
 
@@ -2082,13 +2163,13 @@ class Ota(Cluster):
 
     class ServerCommandDefs(BaseCommandDefs):
         query_next_image: Final = ZCLCommandDef(
-            id=0x01, schema=QueryNextImageCommand, direction=False
+            id=0x01, schema=QueryNextImageCommand, direction=Direction.Client_to_Server
         )
         image_block: Final = ZCLCommandDef(
-            id=0x03, schema=ImageBlockCommand, direction=False
+            id=0x03, schema=ImageBlockCommand, direction=Direction.Client_to_Server
         )
         image_page: Final = ZCLCommandDef(
-            id=0x04, schema=ImagePageCommand, direction=False
+            id=0x04, schema=ImagePageCommand, direction=Direction.Client_to_Server
         )
         upgrade_end: Final = ZCLCommandDef(
             id=0x06,
@@ -2098,7 +2179,7 @@ class Ota(Cluster):
                 "image_type": t.uint16_t,
                 "file_version": t.uint32_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         query_specific_file: Final = ZCLCommandDef(
             id=0x08,
@@ -2109,12 +2190,12 @@ class Ota(Cluster):
                 "file_version": t.uint32_t,
                 "current_zigbee_stack_version": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
         image_notify: Final = ZCLCommandDef(
-            id=0x00, schema=ImageNotifyCommand, direction=False
+            id=0x00, schema=ImageNotifyCommand, direction=Direction.Client_to_Server
         )
         query_next_image_response: Final = ZCLCommandDef(
             id=0x02,
@@ -2125,12 +2206,12 @@ class Ota(Cluster):
                 "file_version?": t.uint32_t,
                 "image_size?": t.uint32_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         image_block_response: Final = ZCLCommandDef(
             id=0x05,
             schema=ImageBlockResponseCommand,
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         upgrade_end_response: Final = ZCLCommandDef(
             id=0x07,
@@ -2141,7 +2222,7 @@ class Ota(Cluster):
                 "current_time": t.UTCTime,
                 "upgrade_time": t.UTCTime,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         query_specific_file_response: Final = ZCLCommandDef(
             id=0x09,
@@ -2152,7 +2233,7 @@ class Ota(Cluster):
                 "file_version?": t.uint32_t,
                 "image_size?": t.uint32_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
 
     def handle_cluster_request(
@@ -2190,17 +2271,12 @@ class Ota(Cluster):
         )
 
         device = self.endpoint.device
-        img = await device.application.ota.get_ota_image(device, cmd)
+        images_result = await device.application.ota.get_ota_images(device, cmd)
 
-        if img is None:
-            self.debug("No OTA image is available")
-            return
-
-        # send an event to listener(s) to let them know that an image is available
         device.listener_event(
-            "device_ota_update_available",
-            img,
-            cmd.current_file_version,
+            "device_ota_image_query_result",
+            images_result,
+            cmd,
         )
 
     async def _handle_image_block_req(self, hdr, cmd):
@@ -2257,10 +2333,12 @@ class PowerProfile(Cluster):
 
     class ServerCommandDefs(BaseCommandDefs):
         power_profile_request: Final = ZCLCommandDef(
-            id=0x00, schema={"power_profile_id": t.uint8_t}, direction=False
+            id=0x00,
+            schema={"power_profile_id": t.uint8_t},
+            direction=Direction.Client_to_Server,
         )
         power_profile_state_request: Final = ZCLCommandDef(
-            id=0x01, schema={}, direction=False
+            id=0x01, schema={}, direction=Direction.Client_to_Server
         )
         get_power_profile_price_response: Final = ZCLCommandDef(
             id=0x02,
@@ -2270,7 +2348,7 @@ class PowerProfile(Cluster):
                 "price": t.uint32_t,
                 "price_trailing_digit": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         get_overall_schedule_price_response: Final = ZCLCommandDef(
             id=0x03,
@@ -2279,7 +2357,7 @@ class PowerProfile(Cluster):
                 "price": t.uint32_t,
                 "price_trailing_digit": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         energy_phases_schedule_notification: Final = ZCLCommandDef(
             id=0x04,
@@ -2287,7 +2365,7 @@ class PowerProfile(Cluster):
                 "power_profile_id": t.uint8_t,
                 "scheduled_phases": t.LVList[ScheduleRecord],
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         energy_phases_schedule_response: Final = ZCLCommandDef(
             id=0x05,
@@ -2295,17 +2373,17 @@ class PowerProfile(Cluster):
                 "power_profile_id": t.uint8_t,
                 "scheduled_phases": t.LVList[ScheduleRecord],
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         power_profile_schedule_constraints_request: Final = ZCLCommandDef(
             id=0x06,
             schema={"power_profile_id": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         energy_phases_schedule_state_request: Final = ZCLCommandDef(
             id=0x07,
             schema={"power_profile_id": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_power_profile_price_extended_response: Final = ZCLCommandDef(
             id=0x08,
@@ -2315,7 +2393,7 @@ class PowerProfile(Cluster):
                 "price": t.uint32_t,
                 "price_trailing_digit": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
@@ -2326,7 +2404,7 @@ class PowerProfile(Cluster):
                 "power_profile_id": t.uint8_t,
                 "transfer_phases": t.LVList[PowerProfilePhase],
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         power_profile_response: Final = ZCLCommandDef(
             id=0x01,
@@ -2335,28 +2413,30 @@ class PowerProfile(Cluster):
                 "power_profile_id": t.uint8_t,
                 "transfer_phases": t.LVList[PowerProfilePhase],
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         power_profile_state_response: Final = ZCLCommandDef(
             id=0x02,
             schema={"power_profiles": t.LVList[PowerProfileType]},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         get_power_profile_price: Final = ZCLCommandDef(
-            id=0x03, schema={"power_profile_id": t.uint8_t}, direction=False
+            id=0x03,
+            schema={"power_profile_id": t.uint8_t},
+            direction=Direction.Client_to_Server,
         )
         power_profile_state_notification: Final = ZCLCommandDef(
             id=0x04,
             schema={"power_profiles": t.LVList[PowerProfileType]},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         get_overall_schedule_price: Final = ZCLCommandDef(
-            id=0x05, schema={}, direction=False
+            id=0x05, schema={}, direction=Direction.Client_to_Server
         )
         energy_phases_schedule_request: Final = ZCLCommandDef(
             id=0x06,
             schema={"power_profile_id": t.uint8_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         energy_phases_schedule_state_response: Final = ZCLCommandDef(
             id=0x07,
@@ -2364,7 +2444,7 @@ class PowerProfile(Cluster):
                 "power_profile_id": t.uint8_t,
                 "num_scheduled_energy_phases": t.uint8_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         energy_phases_schedule_state_notification: Final = ZCLCommandDef(
             id=0x08,
@@ -2372,7 +2452,7 @@ class PowerProfile(Cluster):
                 "power_profile_id": t.uint8_t,
                 "num_scheduled_energy_phases": t.uint8_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         power_profile_schedule_constraints_notification: Final = ZCLCommandDef(
             id=0x09,
@@ -2381,7 +2461,7 @@ class PowerProfile(Cluster):
                 "start_after": t.uint16_t,
                 "stop_before": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
         power_profile_schedule_constraints_response: Final = ZCLCommandDef(
             id=0x0A,
@@ -2390,7 +2470,7 @@ class PowerProfile(Cluster):
                 "start_after": t.uint16_t,
                 "stop_before": t.uint16_t,
             },
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
         get_power_profile_price_extended: Final = ZCLCommandDef(
             id=0x0B,
@@ -2399,7 +2479,7 @@ class PowerProfile(Cluster):
                 "power_profile_id": t.uint8_t,
                 "power_profile_start_time?": t.uint16_t,
             },
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
 
@@ -2453,20 +2533,26 @@ class PollControl(Cluster):
         checkin_response: Final = ZCLCommandDef(
             id=0x00,
             schema={"start_fast_polling": t.Bool, "fast_poll_timeout": t.uint16_t},
-            direction=True,
+            direction=Direction.Server_to_Client,
         )
-        fast_poll_stop: Final = ZCLCommandDef(id=0x01, schema={}, direction=False)
+        fast_poll_stop: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
         set_long_poll_interval: Final = ZCLCommandDef(
-            id=0x02, schema={"new_long_poll_interval": t.uint32_t}, direction=False
+            id=0x02,
+            schema={"new_long_poll_interval": t.uint32_t},
+            direction=Direction.Client_to_Server,
         )
         set_short_poll_interval: Final = ZCLCommandDef(
             id=0x03,
             schema={"new_short_poll_interval": t.uint16_t},
-            direction=False,
+            direction=Direction.Client_to_Server,
         )
 
     class ClientCommandDefs(BaseCommandDefs):
-        checkin: Final = ZCLCommandDef(id=0x0000, schema={}, direction=False)
+        checkin: Final = ZCLCommandDef(
+            id=0x0000, schema={}, direction=Direction.Client_to_Server
+        )
 
 
 class GreenPowerProxy(Cluster):
