@@ -9,6 +9,7 @@ from zigpy.zcl import Cluster
 from zigpy.zcl.foundation import (
     BaseAttributeDefs,
     BaseCommandDefs,
+    Direction,
     ZCLAttributeDef,
     ZCLCommandDef,
 )
@@ -29,14 +30,16 @@ class GenericTunnel(Cluster):
         protocol_addr: Final = ZCLAttributeDef(id=0x0003, type=t.LVBytes)
 
     class ServerCommandDefs(BaseCommandDefs):
-        match_protocol_addr: Final = ZCLCommandDef(id=0x00, schema={}, direction=False)
+        match_protocol_addr: Final = ZCLCommandDef(
+            id=0x00, schema={}, direction=Direction.Client_to_Server
+        )
 
     class ClientCommandDefs(BaseCommandDefs):
         match_protocol_addr_response: Final = ZCLCommandDef(
-            id=0x00, schema={}, direction=True
+            id=0x00, schema={}, direction=Direction.Server_to_Client
         )
         advertise_protocol_address: Final = ZCLCommandDef(
-            id=0x01, schema={}, direction=False
+            id=0x01, schema={}, direction=Direction.Client_to_Server
         )
 
 
@@ -46,7 +49,7 @@ class BacnetProtocolTunnel(Cluster):
 
     class ServerCommandDefs(BaseCommandDefs):
         transfer_npdu: Final = ZCLCommandDef(
-            id=0x00, schema={"npdu": t.LVBytes}, direction=False
+            id=0x00, schema={"npdu": t.LVBytes}, direction=Direction.Client_to_Server
         )
 
 
@@ -83,10 +86,18 @@ class AnalogInputExtended(Cluster):
         # integer, time of day, or structure of (date, time of day))
 
     class ServerCommandDefs(BaseCommandDefs):
-        transfer_apdu: Final = ZCLCommandDef(id=0x00, schema={}, direction=False)
-        connect_req: Final = ZCLCommandDef(id=0x01, schema={}, direction=False)
-        disconnect_req: Final = ZCLCommandDef(id=0x02, schema={}, direction=False)
-        connect_status_noti: Final = ZCLCommandDef(id=0x03, schema={}, direction=False)
+        transfer_apdu: Final = ZCLCommandDef(
+            id=0x00, schema={}, direction=Direction.Client_to_Server
+        )
+        connect_req: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
+        disconnect_req: Final = ZCLCommandDef(
+            id=0x02, schema={}, direction=Direction.Client_to_Server
+        )
+        connect_status_noti: Final = ZCLCommandDef(
+            id=0x03, schema={}, direction=Direction.Client_to_Server
+        )
 
 
 class AnalogOutputRegular(Cluster):
