@@ -7,6 +7,7 @@ from zigpy.zcl import Cluster
 from zigpy.zcl.foundation import (
     BaseAttributeDefs,
     BaseCommandDefs,
+    Direction,
     ZCLAttributeDef,
     ZCLCommandDef,
 )
@@ -419,24 +420,44 @@ class Metering(Cluster):
         )
 
     class ServerCommandDefs(BaseCommandDefs):
-        get_profile: Final = ZCLCommandDef(id=0x00, schema={}, direction=False)
-        req_mirror: Final = ZCLCommandDef(id=0x01, schema={}, direction=False)
-        mirror_rem: Final = ZCLCommandDef(id=0x02, schema={}, direction=False)
-        req_fast_poll_mode: Final = ZCLCommandDef(id=0x03, schema={}, direction=False)
-        get_snapshot: Final = ZCLCommandDef(id=0x04, schema={}, direction=False)
-        take_snapshot: Final = ZCLCommandDef(id=0x05, schema={}, direction=False)
+        get_profile: Final = ZCLCommandDef(
+            id=0x00, schema={}, direction=Direction.Client_to_Server
+        )
+        req_mirror: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Client_to_Server
+        )
+        mirror_rem: Final = ZCLCommandDef(
+            id=0x02, schema={}, direction=Direction.Client_to_Server
+        )
+        req_fast_poll_mode: Final = ZCLCommandDef(
+            id=0x03, schema={}, direction=Direction.Client_to_Server
+        )
+        get_snapshot: Final = ZCLCommandDef(
+            id=0x04, schema={}, direction=Direction.Client_to_Server
+        )
+        take_snapshot: Final = ZCLCommandDef(
+            id=0x05, schema={}, direction=Direction.Client_to_Server
+        )
         mirror_report_attr_response: Final = ZCLCommandDef(
-            id=0x06, schema={}, direction=True
+            id=0x06, schema={}, direction=Direction.Server_to_Client
         )
 
     class ClientCommandDefs(BaseCommandDefs):
-        get_profile_response: Final = ZCLCommandDef(id=0x00, schema={}, direction=True)
-        req_mirror_response: Final = ZCLCommandDef(id=0x01, schema={}, direction=True)
-        mirror_rem_response: Final = ZCLCommandDef(id=0x02, schema={}, direction=True)
-        req_fast_poll_mode_response: Final = ZCLCommandDef(
-            id=0x03, schema={}, direction=True
+        get_profile_response: Final = ZCLCommandDef(
+            id=0x00, schema={}, direction=Direction.Server_to_Client
         )
-        get_snapshot_response: Final = ZCLCommandDef(id=0x04, schema={}, direction=True)
+        req_mirror_response: Final = ZCLCommandDef(
+            id=0x01, schema={}, direction=Direction.Server_to_Client
+        )
+        mirror_rem_response: Final = ZCLCommandDef(
+            id=0x02, schema={}, direction=Direction.Server_to_Client
+        )
+        req_fast_poll_mode_response: Final = ZCLCommandDef(
+            id=0x03, schema={}, direction=Direction.Server_to_Client
+        )
+        get_snapshot_response: Final = ZCLCommandDef(
+            id=0x04, schema={}, direction=Direction.Server_to_Client
+        )
 
 
 class Messaging(Cluster):
