@@ -4,11 +4,14 @@ import dataclasses
 import enum
 import functools
 import keyword
+import logging
 import typing
 
 from typing_extensions import Self
 
 import zigpy.types as t
+
+_LOGGER = logging.getLogger(__name__)
 
 
 def _hex_uint16_repr(v: int) -> str:
@@ -69,16 +72,18 @@ class Status(t.enum8):
         return status
 
 
-class Analog:
-    pass
+class DataClass(enum.Enum):
+    Null = 0
+    Analog = 1
+    Discrete = 2
+    Composite = 3
 
 
-class Discrete:
-    pass
-
-
-class Null:
-    pass
+# TODO: Backwards compatibility, remove later
+Null = DataClass.Null
+Analog = DataClass.Analog
+Discrete = DataClass.Discrete
+Composite = DataClass.Composite
 
 
 class Unknown(t.NoData):
@@ -106,7 +111,7 @@ class TypeValue:
     @classmethod
     def deserialize(cls, data: bytes) -> tuple[TypeValue, bytes]:
         data_type, data = t.uint8_t.deserialize(data)
-        python_type = DATA_TYPES[data_type][1]
+        python_type = DataType.from_type_id(data_type).python_type
         value, data = python_type.deserialize(data)
 
         return cls(type=data_type, value=value), data
@@ -123,7 +128,7 @@ class TypedCollection(TypeValue):
     @classmethod
     def deserialize(cls, data):
         data_type, data = t.uint8_t.deserialize(data)
-        python_type = DATA_TYPES[data_type][1]
+        python_type = DataType.from_type_id(data_type).python_type
         values, data = t.LVList[python_type, t.uint16_t].deserialize(data)
 
         return cls(type=data_type, value=values), data
@@ -141,108 +146,559 @@ class Set(TypedCollection):
     pass  # ToDo: Make this a real set?
 
 
-class DataTypes(dict):
-    """DataTypes container."""
-
-    def __init__(
-        self,
-        data_types: dict[
-            int,
-            tuple[
-                str,
-                typing.Any,
-                typing.Literal[Null, Discrete, Analog, None],
-            ],
-        ],
-    ) -> None:
-        super().__init__(data_types)
-        self._idx_by_class = {
-            _type: type_id for type_id, (name, _type, ad) in self.items()
-        }
-
-    def pytype_to_datatype_id(self, python_type: typing.Any) -> int:
-        """Return Zigbee Datatype ID for a give python type."""
-
-        # We return the most specific parent class
-        for cls in python_type.__mro__:
-            if cls in self._idx_by_class:
-                return self._idx_by_class[cls]
-
-        return 0xFF
-
-
 class ZCLStructure(t.LVList, item_type=TypeValue, length_type=t.uint16_t):
     """ZCL Structure data type."""
 
 
-DATA_TYPES = DataTypes(
-    {
-        0x00: ("No data", t.NoData, Null),
-        0x08: ("General", t.data8, Discrete),
-        0x09: ("General", t.data16, Discrete),
-        0x0A: ("General", t.data24, Discrete),
-        0x0B: ("General", t.data32, Discrete),
-        0x0C: ("General", t.data40, Discrete),
-        0x0D: ("General", t.data48, Discrete),
-        0x0E: ("General", t.data56, Discrete),
-        0x0F: ("General", t.data64, Discrete),
-        0x10: ("Boolean", t.Bool, Discrete),
-        0x18: ("Bitmap", t.bitmap8, Discrete),
-        0x19: ("Bitmap", t.bitmap16, Discrete),
-        0x1A: ("Bitmap", t.bitmap24, Discrete),
-        0x1B: ("Bitmap", t.bitmap32, Discrete),
-        0x1C: ("Bitmap", t.bitmap40, Discrete),
-        0x1D: ("Bitmap", t.bitmap48, Discrete),
-        0x1E: ("Bitmap", t.bitmap56, Discrete),
-        0x1F: ("Bitmap", t.bitmap64, Discrete),
-        0x20: ("Unsigned Integer", t.uint8_t, Analog),
-        0x21: ("Unsigned Integer", t.uint16_t, Analog),
-        0x22: ("Unsigned Integer", t.uint24_t, Analog),
-        0x23: ("Unsigned Integer", t.uint32_t, Analog),
-        0x24: ("Unsigned Integer", t.uint40_t, Analog),
-        0x25: ("Unsigned Integer", t.uint48_t, Analog),
-        0x26: ("Unsigned Integer", t.uint56_t, Analog),
-        0x27: ("Unsigned Integer", t.uint64_t, Analog),
-        0x28: ("Signed Integer", t.int8s, Analog),
-        0x29: ("Signed Integer", t.int16s, Analog),
-        0x2A: ("Signed Integer", t.int24s, Analog),
-        0x2B: ("Signed Integer", t.int32s, Analog),
-        0x2C: ("Signed Integer", t.int40s, Analog),
-        0x2D: ("Signed Integer", t.int48s, Analog),
-        0x2E: ("Signed Integer", t.int56s, Analog),
-        0x2F: ("Signed Integer", t.int64s, Analog),
-        0x30: ("Enumeration", t.enum8, Discrete),
-        0x31: ("Enumeration", t.enum16, Discrete),
-        0x38: ("Floating point", t.Half, Analog),
-        0x39: ("Floating point", t.Single, Analog),
-        0x3A: ("Floating point", t.Double, Analog),
-        0x41: ("Octet string", t.LVBytes, Discrete),
-        0x42: ("Character string", t.CharacterString, Discrete),
-        0x43: ("Long octet string", t.LongOctetString, Discrete),
-        0x44: ("Long character string", t.LongCharacterString, Discrete),
-        0x48: ("Array", Array, Discrete),
-        0x4C: ("Structure", ZCLStructure, Discrete),
-        0x50: ("Set", Set, Discrete),
-        0x51: ("Bag", Bag, Discrete),
-        0xE0: ("Time of day", t.TimeOfDay, Analog),
-        0xE1: ("Date", t.Date, Analog),
-        0xE2: ("UTCTime", t.UTCTime, Analog),
-        0xE8: ("Cluster ID", t.ClusterId, Discrete),
-        0xE9: ("Attribute ID", t.AttributeId, Discrete),
-        0xEA: ("BACNet OID", t.BACNetOid, Discrete),
-        0xF0: ("IEEE address", t.EUI64, Discrete),
-        0xF1: ("128-bit security key", t.KeyData, Discrete),
-        0xFF: ("Unknown", Unknown, None),
-    }
-)
+class DataTypeId(t.enum8):
+    unk = 0xFF
+    nodata = 0x00
+    data8 = 0x08
+    data16 = 0x09
+    data24 = 0x0A
+    data32 = 0x0B
+    data40 = 0x0C
+    data48 = 0x0D
+    data56 = 0x0E
+    data64 = 0x0F
+    bool_ = 0x10
+    map8 = 0x18
+    map16 = 0x19
+    map24 = 0x1A
+    map32 = 0x1B
+    map40 = 0x1C
+    map48 = 0x1D
+    map56 = 0x1E
+    map64 = 0x1F
+    uint8 = 0x20
+    uint16 = 0x21
+    uint24 = 0x22
+    uint32 = 0x23
+    uint40 = 0x24
+    uint48 = 0x25
+    uint56 = 0x26
+    uint64 = 0x27
+    int8 = 0x28
+    int16 = 0x29
+    int24 = 0x2A
+    int32 = 0x2B
+    int40 = 0x2C
+    int48 = 0x2D
+    int56 = 0x2E
+    int64 = 0x2F
+    enum8 = 0x30
+    enum16 = 0x31
+    semi = 0x38
+    single = 0x39
+    double = 0x3A
+    octstr = 0x41
+    string = 0x42
+    octstr16 = 0x43
+    string16 = 0x44
+    array = 0x48
+    struct = 0x4C
+    set = 0x50
+    bag = 0x51
+    ToD = 0xE0
+    date = 0xE1
+    UTC = 0xE2
+    clusterId = 0xE8  # noqa: N815
+    attribId = 0xE9  # noqa: N815
+    bacOID = 0xEA  # noqa: N815
+    EUI64 = 0xF0
+    key128 = 0xF1
 
 
-class ReadAttributeRecord(t.Struct):
+@dataclasses.dataclass(frozen=True)
+class DataTypeInfo:
+    type_id: DataTypeId
+    python_type: type
+    type_class: DataClass
+    description: str
+    non_value: typing.Any | None
+
+
+class DataType(DataTypeInfo, enum.Enum):
+    unk = (
+        DataTypeId.unk,
+        Unknown,
+        DataClass.Null,
+        "Unknown",
+        None,
+    )
+    nodata = (
+        DataTypeId.nodata,
+        t.NoData,
+        DataClass.Null,
+        "No data",
+        None,
+    )
+    data8 = (
+        DataTypeId.data8,
+        t.data8,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    data16 = (
+        DataTypeId.data16,
+        t.data16,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    data24 = (
+        DataTypeId.data24,
+        t.data24,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    data32 = (
+        DataTypeId.data32,
+        t.data32,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    data40 = (
+        DataTypeId.data40,
+        t.data40,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    data48 = (
+        DataTypeId.data48,
+        t.data48,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    data56 = (
+        DataTypeId.data56,
+        t.data56,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    data64 = (
+        DataTypeId.data64,
+        t.data64,
+        DataClass.Discrete,
+        "General",
+        None,
+    )
+    bool_ = (
+        DataTypeId.bool_,
+        t.Bool,
+        DataClass.Discrete,
+        "Boolean",
+        t.Bool(0xFF),
+    )
+    map8 = (
+        DataTypeId.map8,
+        t.bitmap8,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    map16 = (
+        DataTypeId.map16,
+        t.bitmap16,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    map24 = (
+        DataTypeId.map24,
+        t.bitmap24,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    map32 = (
+        DataTypeId.map32,
+        t.bitmap32,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    map40 = (
+        DataTypeId.map40,
+        t.bitmap40,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    map48 = (
+        DataTypeId.map48,
+        t.bitmap48,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    map56 = (
+        DataTypeId.map56,
+        t.bitmap56,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    map64 = (
+        DataTypeId.map64,
+        t.bitmap64,
+        DataClass.Discrete,
+        "Bitmap",
+        None,
+    )
+    uint8 = (
+        DataTypeId.uint8,
+        t.uint8_t,
+        DataClass.Analog,
+        "Unsigned 8-bit integer",
+        t.uint8_t(0xFF),
+    )
+    uint16 = (
+        DataTypeId.uint16,
+        t.uint16_t,
+        DataClass.Analog,
+        "Unsigned 16-bit integer",
+        t.uint16_t(0xFFFF),
+    )
+    uint24 = (
+        DataTypeId.uint24,
+        t.uint24_t,
+        DataClass.Analog,
+        "Unsigned 24-bit integer",
+        t.uint24_t(0xFFFFFF),
+    )
+    uint32 = (
+        DataTypeId.uint32,
+        t.uint32_t,
+        DataClass.Analog,
+        "Unsigned 32-bit integer",
+        t.uint32_t(0xFFFFFFFF),
+    )
+    uint40 = (
+        DataTypeId.uint40,
+        t.uint40_t,
+        DataClass.Analog,
+        "Unsigned 40-bit integer",
+        t.uint40_t(0xFFFFFFFFFF),
+    )
+    uint48 = (
+        DataTypeId.uint48,
+        t.uint48_t,
+        DataClass.Analog,
+        "Unsigned 48-bit integer",
+        t.uint48_t(0xFFFFFFFFFFFF),
+    )
+    uint56 = (
+        DataTypeId.uint56,
+        t.uint56_t,
+        DataClass.Analog,
+        "Unsigned 56-bit integer",
+        t.uint56_t(0xFFFFFFFFFFFFFF),
+    )
+    uint64 = (
+        DataTypeId.uint64,
+        t.uint64_t,
+        DataClass.Analog,
+        "Unsigned 64-bit integer",
+        t.uint64_t(0xFFFFFFFFFFFFFF),
+    )
+    int8 = (
+        DataTypeId.int8,
+        t.int8s,
+        DataClass.Analog,
+        "Signed 8-bit integer",
+        t.int8s(-0x80),
+    )
+    int16 = (
+        DataTypeId.int16,
+        t.int16s,
+        DataClass.Analog,
+        "Signed 16-bit integer",
+        t.int16s(-0x8000),
+    )
+    int24 = (
+        DataTypeId.int24,
+        t.int24s,
+        DataClass.Analog,
+        "Signed 24-bit integer",
+        t.int24s(-0x800000),
+    )
+    int32 = (
+        DataTypeId.int32,
+        t.int32s,
+        DataClass.Analog,
+        "Signed 32-bit integer",
+        t.int32s(-0x80000000),
+    )
+    int40 = (
+        DataTypeId.int40,
+        t.int40s,
+        DataClass.Analog,
+        "Signed 40-bit integer",
+        t.int40s(-0x8000000000),
+    )
+    int48 = (
+        DataTypeId.int48,
+        t.int48s,
+        DataClass.Analog,
+        "Signed 48-bit integer",
+        t.int48s(-0x800000000000),
+    )
+    int56 = (
+        DataTypeId.int56,
+        t.int56s,
+        DataClass.Analog,
+        "Signed 56-bit integer",
+        t.int56s(-0x80000000000000),
+    )
+    int64 = (
+        DataTypeId.int64,
+        t.int64s,
+        DataClass.Analog,
+        "Signed 64-bit integer",
+        t.int64s(-0x80000000000000),
+    )
+    enum8 = (
+        DataTypeId.enum8,
+        t.enum8,
+        DataClass.Discrete,
+        "8-bit enumeration",
+        t.enum8(0xFF),
+    )
+    enum16 = (
+        DataTypeId.enum16,
+        t.enum16,
+        DataClass.Discrete,
+        "16-bit enumeration",
+        t.enum16(0xFF),
+    )
+    semi = (
+        DataTypeId.semi,
+        t.Half,
+        DataClass.Analog,
+        "Semi-precision",
+        None,
+    )
+    single = (
+        DataTypeId.single,
+        t.Single,
+        DataClass.Analog,
+        "Single precision",
+        None,
+    )
+    double = (
+        DataTypeId.double,
+        t.Double,
+        DataClass.Analog,
+        "Double precision",
+        None,
+    )
+    octstr = (
+        DataTypeId.octstr,
+        t.LVBytes,
+        DataClass.Discrete,
+        "Octet string",
+        None,
+    )
+    string = (
+        DataTypeId.string,
+        t.CharacterString,
+        DataClass.Discrete,
+        "Character string",
+        None,
+    )
+    octstr16 = (
+        DataTypeId.octstr16,
+        t.LongOctetString,
+        DataClass.Discrete,
+        "Long octet string",
+        None,
+    )
+    string16 = (
+        DataTypeId.string16,
+        t.LongCharacterString,
+        DataClass.Discrete,
+        "Long character string",
+        None,
+    )
+    array = (
+        DataTypeId.array,
+        Array,
+        DataClass.Discrete,
+        "Array",
+        None,
+    )
+    struct = (
+        DataTypeId.struct,
+        ZCLStructure,
+        DataClass.Discrete,
+        "Structure",
+        None,
+    )
+    set = (
+        DataTypeId.set,
+        Set,
+        DataClass.Discrete,
+        "Set",
+        None,
+    )
+    bag = (
+        DataTypeId.bag,
+        Bag,
+        DataClass.Discrete,
+        "Bag",
+        None,
+    )
+    ToD = (
+        DataTypeId.ToD,
+        t.TimeOfDay,
+        DataClass.Analog,
+        "Time of day",
+        t.TimeOfDay(hours=0xFF, minutes=0xFF, seconds=0xFF, hundredths=0xFF),
+    )
+    date = (
+        DataTypeId.date,
+        t.Date,
+        DataClass.Analog,
+        "Date",
+        t.Date(years_since_1900=0xFF, month=0xFF, day=0xFF, day_of_week=0xFF),
+    )
+    UTC = (
+        DataTypeId.UTC,
+        t.UTCTime,
+        DataClass.Analog,
+        "UTCTime",
+        t.UTCTime(0xFFFFFFFF),
+    )
+    clusterId = (  # noqa: N815
+        DataTypeId.clusterId,
+        t.ClusterId,
+        DataClass.Discrete,
+        "Cluster ID",
+        t.ClusterId(0xFFFF),
+    )
+    attribId = (  # noqa: N815
+        DataTypeId.attribId,
+        t.AttributeId,
+        DataClass.Discrete,
+        "Attribute ID",
+        t.AttributeId(0xFFFF),
+    )
+    bacOID = (  # noqa: N815
+        DataTypeId.bacOID,
+        t.BACNetOid,
+        DataClass.Discrete,
+        "BACNet OID",
+        t.BACNetOid(0xFFFFFFFF),
+    )
+    EUI64 = (
+        DataTypeId.EUI64,
+        t.EUI64,
+        DataClass.Discrete,
+        "IEEE address",
+        t.EUI64.convert("FF:FF:FF:FF:FF:FF:FF:FF"),
+    )
+    key128 = (
+        DataTypeId.key128,
+        t.KeyData,
+        DataClass.Discrete,
+        "128-bit security key",
+        t.KeyData.convert("FF:FF:FF:FF:FF:FF:FF:FF:FF:FF:FF:FF:FF:FF:FF:FF"),
+    )
+
+    @classmethod
+    @functools.cache
+    def _python_type_index(cls: type[Self]) -> dict[type, Self]:  # noqa: N805
+        return {d.python_type: d for d in cls}
+
+    @classmethod
+    def from_python_type(cls: type[Self], python_type: type) -> Self:
+        """Return Zigbee Datatype ID for a give python type."""
+        python_type_index = cls._python_type_index()
+
+        # We return the most specific parent class
+        for parent_cls in python_type.__mro__:
+            if parent_cls in python_type_index:
+                return python_type_index[parent_cls]
+
+        return cls.unk
+
+    @classmethod
+    @functools.cache
+    def _data_type_index(cls: type[Self]) -> dict[type, Self]:  # noqa: N805
+        return {d.type_id: d for d in cls}
+
+    @classmethod
+    def from_type_id(cls: type[Self], type_id: DataTypeId) -> Self:
+        return cls._data_type_index()[type_id]
+
+
+@dataclasses.dataclass()
+class ReadAttributeRecord:
     """Read Attribute Record."""
 
-    attrid: t.uint16_t = t.StructField(repr=_hex_uint16_repr)
+    attrid: t.uint16_t
     status: Status
-    value: TypeValue = t.StructField(requires=lambda s: s.status == Status.SUCCESS)
+    value: TypeValue | Array | Bag | Set | None
+
+    def __init__(
+        self,
+        attrid: t.uint16_t | Self = t.uint16_t(0x0000),
+        status: Status = Status.SUCCESS,
+        value: TypeValue | Array | Bag | Set | None = None,
+    ) -> None:
+        if isinstance(attrid, self.__class__):
+            # "Copy constructor"
+            self.attrid = attrid.attrid
+            self.status = attrid.status
+            self.value = attrid.value
+            return
+
+        self.attrid = t.uint16_t(attrid)
+        self.status = Status(status)
+        self.value = value
+
+    @classmethod
+    def deserialize(cls, data: bytes) -> tuple[Self, bytes]:
+        attrid, data = t.uint16_t.deserialize(data)
+        status, data = Status.deserialize(data)
+        value = None
+
+        if status == Status.SUCCESS:
+            type_id, data = DataTypeId.deserialize(data)
+
+            # Arrays, Sets, and Bags are treated differently
+            if type_id in (DataTypeId.array, DataTypeId.set, DataTypeId.bag):
+                value, data = DataType.from_type_id(type_id).python_type.deserialize(
+                    data
+                )
+            else:
+                value, data = TypeValue.deserialize(type_id.serialize() + data)
+
+        return cls(attrid=attrid, status=status, value=value), data
+
+    def serialize(self) -> bytes:
+        data = self.attrid.serialize()
+        data += self.status.serialize()
+
+        if self.status == Status.SUCCESS:
+            assert self.value is not None
+
+            if isinstance(self.value, (Array, Set, Bag)):
+                data += (
+                    DataType.from_python_type(type(self.value)).type_id.serialize()
+                    + self.value.serialize()
+                )
+            else:
+                data += self.value.serialize()
+
+        return data
 
 
 class Attribute(t.Struct):
@@ -304,7 +760,7 @@ class AttributeReportingConfig:
             if self.direction == ReportingDirection.ReceiveReports:
                 self.timeout: int = other.timeout
                 return
-            self.datatype: int = other.datatype
+            self.datatype: DataTypeId = other.datatype
             self.min_interval: int = other.min_interval
             self.max_interval: int = other.max_interval
             self.reportable_change: int = other.reportable_change
@@ -322,10 +778,17 @@ class AttributeReportingConfig:
             r += t.uint8_t(self.datatype).serialize()
             r += t.uint16_t(self.min_interval).serialize()
             r += t.uint16_t(self.max_interval).serialize()
-            datatype = DATA_TYPES.get(self.datatype, None)
-            if datatype and datatype[2] is Analog:
-                datatype = datatype[1]
-                r += datatype(self.reportable_change).serialize()
+
+            try:
+                data_type = DataType.from_type_id(self.datatype)
+            except KeyError:
+                _LOGGER.warning(
+                    "Unknown ZCL type %d, not setting reportable change", self.datatype
+                )
+            else:
+                if data_type.type_class is Analog:
+                    r += data_type.python_type(self.reportable_change).serialize()
+
         return r
 
     @classmethod
@@ -346,11 +809,20 @@ class AttributeReportingConfig:
         else:
             # Notifying that I will report things to you
             self.datatype, data = t.uint8_t.deserialize(data)
-            datatype = DATA_TYPES[self.datatype]
             self.min_interval, data = t.uint16_t.deserialize(data)
             self.max_interval, data = t.uint16_t.deserialize(data)
-            if datatype[2] is Analog:
-                self.reportable_change, data = datatype[1].deserialize(data)
+
+            try:
+                data_type = DataType.from_type_id(self.datatype)
+            except KeyError:
+                _LOGGER.warning(
+                    "Unknown ZCL type %d, cannot read reportable change", self.datatype
+                )
+            else:
+                if data_type.type_class is Analog:
+                    self.reportable_change, data = data_type.python_type.deserialize(
+                        data
+                    )
 
         return self, data
 
@@ -770,12 +1242,9 @@ ZCLAttributeAccess._names = {
 class ZCLAttributeDef(t.BaseDataclassMixin):
     id: t.uint16_t = None
     type: type = None
-    access: ZCLAttributeAccess = dataclasses.field(
-        default=(
-            ZCLAttributeAccess.Read
-            | ZCLAttributeAccess.Write
-            | ZCLAttributeAccess.Report
-        ),
+    zcl_type: DataTypeId = None
+    access: ZCLAttributeAccess = (
+        ZCLAttributeAccess.Read | ZCLAttributeAccess.Write | ZCLAttributeAccess.Report
     )
     mandatory: bool = False
     is_manufacturer_specific: bool = False
@@ -795,6 +1264,11 @@ class ZCLAttributeDef(t.BaseDataclassMixin):
         if isinstance(self.access, str):
             object.__setattr__(self, "access", ZCLAttributeAccess.from_str(self.access))
 
+        if self.zcl_type is None:
+            object.__setattr__(
+                self, "zcl_type", DataType.from_python_type(self.type).type_id
+            )
+
         ensure_valid_name(self.name)
 
     def __repr__(self) -> str:
@@ -803,6 +1277,7 @@ class ZCLAttributeDef(t.BaseDataclassMixin):
             f"id=0x{self.id:04X}, "
             f"name={self.name!r}, "
             f"type={self.type}, "
+            f"zcl_type={self.zcl_type}, "
             f"access={self.access!r}, "
             f"mandatory={self.mandatory!r}, "
             f"is_manufacturer_specific={self.is_manufacturer_specific}"
